feat: Separate pages to be able to select mode
